requirements.txt
flask==2.3.3
langchain==0.3.5
langchain-community==0.3.1
openai==1.68.2
pinecone-client==5.0.1  # Updated from 3.0.0
tiktoken==0.7.0
pymupdf==1.23.22
markdown==3.6
python-dotenv==1.0.1
langchain_pinecone==0.2.0  # Added explicitly
gunicorn==23.0.0